Toxic Memer

+++ events/interactionCreate.js
@@ -3,9 +3,7 @@ module.exports = {
     once: false,
     run: async (interaction, client) => {
         if (interaction.isButton()) {
-            if (interaction.customId == 'delete_eval') {
-                return interaction.message.delete();
-            }
+            if (interaction.customId == 'delete_eval') return interaction.message.delete();
         }
     }
 }
